flask-server/requirements.txt
requests
pymongo[srv] 
scikit-learn
flask
python-dotenv 